Rename ogr.py to file.py. Remove unused import
